repair: extract repair_reader declaration into reader.hh
It's needed to write a unit test for it in the following
commits. No other code changes have been made.
